Drop encoding parameter from json.load
Apparently this is no longer an accepted keyword for json.load(), as
of python 3.9.

This was uncovered in https://bugs.debian.org/973175 where the
google-i18n-address tests failed.  But i haven't been able to find a
reference for the change in python upstream.  Perhaps the encoding
parameter never did anything, but 3.9 became more strict about
unhandled keywords.

It looks like python's json module does bytestream encoding
autodetection anyway so this should be safe to do:
https://bugs.python.org/issue17909